+++ frontend/src/components/visual/WaveAninmation.js
